Splitting up Selenium class into logical enums
