+++ src/main/java/com/coveros/selenified/element/WaitFor.java
@@ -18,7 +18,7 @@
  * under the License.
  */
 
-package com.coveros.selenified.selenium.element;
+package com.coveros.selenified.element;
 
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.StaleElementReferenceException;
@@ -27,7 +27,7 @@ import org.testng.log4testng.Logger;
 
 import com.coveros.selenified.OutputFile;
 import com.coveros.selenified.OutputFile.Result;
-import com.coveros.selenified.selenium.element.Element;
+import com.coveros.selenified.element.Element;
 
 /**
  * WaitFor performs dynamic waits on a particular element, until a particular
