+++ src/test/java/integration/ActionWaitIT.java
@@ -6,10 +6,10 @@ import org.testng.ITestContext;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import com.coveros.selenified.selenium.App;
-import com.coveros.selenified.selenium.Selenium.Locator;
-import com.coveros.selenified.selenium.element.Element;
 import com.coveros.selenified.Selenified;
+import com.coveros.selenified.Locator;
+import com.coveros.selenified.application.App;
+import com.coveros.selenified.element.Element;
 
 public class ActionWaitIT extends Selenified {
