+++ src/test/java/integration/NoBrowserIT.java
@@ -9,10 +9,10 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import com.coveros.selenified.selenium.App;
-import com.coveros.selenified.selenium.Selenium.Browser;
-import com.coveros.selenified.selenium.Selenium.DriverSetup;
 import com.coveros.selenified.Selenified;
+import com.coveros.selenified.Browser;
+import com.coveros.selenified.DriverSetup;
+import com.coveros.selenified.application.App;
 
 public class NoBrowserIT extends Selenified {
