+++ src/test/java/sample/SimpleSampleIT.java
@@ -5,10 +5,10 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
-import com.coveros.selenified.selenium.App;
-import com.coveros.selenified.selenium.Selenium.Locator;
-import com.coveros.selenified.selenium.element.Element;
 import com.coveros.selenified.Selenified;
+import com.coveros.selenified.Locator;
+import com.coveros.selenified.application.App;
+import com.coveros.selenified.element.Element;
 
 public class SimpleSampleIT extends Selenified {
