+++ src/test/java/unit/ExceptionTest.java
@@ -4,8 +4,8 @@ import org.openqa.selenium.remote.DesiredCapabilities;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
+import com.coveros.selenified.Browser;
 import com.coveros.selenified.exceptions.InvalidBrowserException;
-import com.coveros.selenified.selenium.Selenium.Browser;
 import com.coveros.selenified.utilities.TestSetup;
 
 public class ExceptionTest {
