+++ src/test/java/unit/ResponseTest.java
@@ -11,7 +11,7 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import com.coveros.selenified.OutputFile;
-import com.coveros.selenified.selenium.Selenium.Browser;
+import com.coveros.selenified.Browser;
 import com.coveros.selenified.services.Response;
 import com.google.common.base.Charsets;
 import com.google.common.io.Files;
