+++ src/test/java/unit/SeleniumTest.java
@@ -4,9 +4,9 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import com.coveros.selenified.OutputFile.Success;
+import com.coveros.selenified.Browser;
+import com.coveros.selenified.DriverSetup;
 import com.coveros.selenified.exceptions.InvalidBrowserException;
-import com.coveros.selenified.selenium.Selenium.Browser;
-import com.coveros.selenified.selenium.Selenium.DriverSetup;
 
 public class SeleniumTest {
